[Docs]#1 Networkable 별도 파일 생성
