+++ OpenMarket/OpenMarket/Controller/Networkable.swift
@@ -1,20 +1,19 @@
 //
-//  URLSessionManager.swift
+//  Networkable.swift
 //  OpenMarket
 //
-//  Created by 김찬우 on 2021/05/28.
+//  Created by 김찬우 on 2021/06/03.
 //
 
 import Foundation
-class URLSessionManager<T: Decodable> {
-    let clientRequest: GETRequest
-    
-    init(clientRequest: GETRequest){
-        self.clientRequest = clientRequest
-    }
-    
-    func getServerData<T: Decodable>(completionHandler: @escaping (T) -> Void) {
-        URLSession.shared.dataTask(with: clientRequest.urlRequest.url!){ data, response, error in
+
+protocol Networkable {
+    func getServerData<T: Decodable>(url: URL, completionHandler: @escaping (T) -> Void)
+}
+
+extension Networkable {
+    func getServerData<T: Decodable>(url: URL, completionHandler: @escaping (T) -> Void) {
+        URLSession.shared.dataTask(with: url){ data, response, error in
             if error != nil { return }
             
             guard let httpResponse = response as? HTTPURLResponse,
